+++ sdk/keyvault/azure_security_keyvault_certificates/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Removed `CertificateClient::begin_create_certificate()`.
 - Removed `CertificateClient::resume_create_certificate()`.
 - Removed `wait()` function from `Poller<CertificateOperation>`.
+- Changed `PollerOptions::frequency` from `Option<Duration>` to `Duration`.
 
 ### Bugs Fixed